CropImage/requirements.txt
appdirs==1.4.4
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
boto3==1.23.0
botocore==1.26.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
cycler==0.11.0
decorator==5.1.1
dj-database-url==0.5.0
Django==3.2.10
django-heroku==0.3.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
dynaconf==3.1.8
fonttools==4.29.1
h11==0.13.0
hvac==0.11.2
idna==3.3
image==1.5.33
Jinja2==3.1.2
jmespath==1.0.0
kiwisolver==1.3.2
knox==0.1.14
loguru==0.6.0
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.22.2
opencv-python==4.8.0.74
outcome==1.2.0
packaging==21.3
Pillow==9.0.1
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
pyglet==1.5.26
pyOpenSSL==22.0.0
pyparsing==3.0.7
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQt5-stubs==5.15.6.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyshark==0.5.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.3
requests==2.27.1
response==0.3.0
s3transfer==0.5.2
scipy==1.8.0
selenium==4.3.0
shortuuid==1.0.9
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.21.0
trio-websocket==0.9.2
tzdata==2021.5
urllib3==1.26.9
validators==0.19.0
whitenoise==6.1.0
win32-setctime==1.1.0
wsproto==1.1.0
